% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselearner.R
\name{dt.fit}
\alias{dt.fit}
\title{Fit function for decision tree models}
\usage{
dt.fit(dtree, X, y, lr)
}
\arguments{
\item{dtree}{A list containing the decision tree model.}

\item{X}{The input data for fitting.}

\item{y}{The response variable.}

\item{lr}{The learning rate.}
}
\value{
A list containing the updated decision tree model.
}
\description{
This function fits a decision tree model using the given data and learning rate.
}
