% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselearner.R
\name{dt.predict}
\alias{dt.predict}
\title{Predict function for decision tree models}
\usage{
dt.predict(dtree, X)
}
\arguments{
\item{dtree}{A list containing the decision tree model.}

\item{X}{The input data for prediction.}
}
\value{
A vector of predicted response values.
}
\description{
This function predicts the response variable using a decision tree model.
}
