% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grbLMM.R
\name{grbLMM}
\alias{grbLMM}
\title{Generalized Randomized Boosted Linear Mixed Models}
\usage{
grbLMM(
  y,
  X,
  Z,
  id,
  beta.fit = NULL,
  beta.predict = NULL,
  beta.init = NULL,
  beta.keep.all = TRUE,
  m.stop = 500,
  ny = 0.1,
  cv.dat = NULL,
  aic = FALSE,
  ...
)
}
\arguments{
\item{y}{The dependent variable.}

\item{X}{Numeric design matrix for the fixed effects (no intercept column).}

\item{Z}{Numeric design matrix for the random effects, same shape as X with intercepts in the first column.}

\item{id}{Grouping factor.}

\item{beta.fit}{Fit function for fixed effects. Pass in wrappers in form of beta.fit(beta, X, u, ny) for decision tree, lasso, etc. Default NULL means using grbLMM baseline.}

\item{beta.predict}{Predict function for fixed effects. Pass in wrappers in form of beta.predict(beta, X) for decision tree, lasso, etc. Default NULL means using grbLMM baseline.}

\item{beta.init}{Initial value for fixed effects that will be passed into beta.predict and beta.fit during first iteration. Default NULL and can remain unchanged if beta.predict and beta.fit allows NULL input.}

\item{beta.keep.all}{Whether to return all betas in all iterations. Should be disabled if single beta already stores the previous boosting results.}

\item{m.stop}{Given number of total iterations.}

\item{ny}{Learning rate, i.e., step length, of the algorithm.}

\item{cv.dat}{Optional data for evaluation.}

\item{aic}{Whether to compute AICc.}

\item{...}{Additional arguments passed to the fit and predict functions.}
}
\value{
A list containing the model parameters and evaluation metrics.
}
\description{
Fits a boosting algorithm for linear mixed models with optional cross-validation.
}
