% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{k_fold_by_prop}
\alias{k_fold_by_prop}
\title{Split Data into k Folds by Proportion}
\usage{
k_fold_by_prop(ds, prop)
}
\arguments{
\item{ds}{A data frame containing the dataset to be split. It should have a column 'id' indicating the subject identifiers.}

\item{prop}{Proportion of data to be used for each fold (0 < prop < 1).}
}
\value{
A list of length k, where each element is a data frame representing a fold.
}
\description{
This function splits the input data into k folds for cross-validation, ensuring each subject has complete observations in each fold.
}
\examples{
# Split data into folds with each fold containing 20\% of the data
folds <- k_fold_by_prop(dataset, prop = 0.2)

}
\keyword{data}
\keyword{manipulation}
