% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grbLMM.R
\name{predict.grbLMM}
\alias{predict.grbLMM}
\title{Predict Method for Generalized Randomized Boosted Linear Mixed Models}
\usage{
\method{predict}{grbLMM}(model, X, Z, id, beta.predict = NULL)
}
\arguments{
\item{model}{A trained model object returned by \code{\link{grbLMM}}.}

\item{X}{Numeric design matrix for the fixed effects (no intercept column).}

\item{Z}{Numeric design matrix for the random effects, same shape as X with intercepts in the first column.}

\item{id}{Grouping factor.}

\item{beta.predict}{Predict function for fixed effects.}
}
\value{
A vector of predicted response values.
}
\description{
Predicts the response variable for a given model and new data.
}
