% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselearner.R
\name{xgb.fit}
\alias{xgb.fit}
\title{Fit function for XGBoost models}
\usage{
xgb.fit(xgb, X, y, lr, lambda = 1, alpha = 0)
}
\arguments{
\item{xgb}{A list containing the XGBoost model.}

\item{X}{The input data for fitting.}

\item{y}{The response variable.}

\item{lr}{The learning rate.}

\item{lambda}{L2 Regularization parameter on weights (default: 1).}

\item{alpha}{L1 regularization parameter on weights (default: 0).}
}
\value{
A list containing the updated XGBoost model.
}
\description{
This function fits an XGBoost model using the given data, learning rate, lambda, and alpha parameters.
}
