% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselearner.R
\name{xgb.predict}
\alias{xgb.predict}
\title{Predict function for XGBoost models}
\usage{
xgb.predict(xgb, X)
}
\arguments{
\item{xgb}{A list containing the XGBoost model.}

\item{X}{The input data for prediction.}
}
\value{
A vector of predicted response values.
}
\description{
This function predicts the response variable using an XGBoost model.
}
